Moved some files around.  Corrections and enhancements to get setup and setup_drive working.

+++ servers/bc.uchsc.edu/lib/amanda/setup_drive.rb
@@ -6,28 +6,33 @@ module Amanda
   #
   class SetupDrive < Tap::FileTask
     
-    config :space, "/dumps"   # TODO -- check not nil, and does not exist.  the space folder for the dumps
-    config :n_slots, 5        # TODO -- check > 1
+    config :n_tapes, 33       
     config :user, 'amandabackup'
     config :group, 'disk'
 
-    def process(name)
-      sh "mkdir -p #{space}/vtapes"
-      sh "chown #{user}:#{group} #{space}/vtapes"
-      sh "chmod 750 #{space}/vtapes"
+    def process(name, space)
+      
+      raise "already exists: #{space}/vtapes/#{name}" if File.exists?("#{space}/vtapes/#{name}")
+      
+      sh "mkdir -p #{space}/vtapes/#{name}"
+      sh "chown #{user}:#{group} #{space}/vtapes/#{name}"
+      sh "chmod 750 #{space}/vtapes/#{name}"
 
       sh "sudo -u #{user} mkdir -p #{space}/vtapes/#{name}/slots"
 
       FileUtils.cd "#{space}/vtapes/#{name}/slots"
-      1.upto(n_slots) do |slot|
+      1.upto(n_tapes) do |slot|
         sh "sudo -u #{user} mkdir slot#{slot}"
       end
 
       sh "sudo -u #{user} ln -s slot1 data"
 
-      1.upto(n_slots) do |slot|
-        sh "sudo -u #{user} amlabel #{name} #{name}-#{slot} slot #{slot}"
+      1.upto(n_tapes) do |slot|
+        sh "sudo -u #{user} amlabel #{name} #{name}-#{slot < 10 ? '0' : ''}#{slot} slot #{slot}"
       end
+      
+      sh "sudo -u #{user} amtape #{name} reset"
+      sh "sudo -u #{user} amcheck #{name}"
     end
     
   end
